Remove old ApiClient
