+++ admin/src/main/scala/net/wiringbits/API.scala
@@ -22,7 +22,7 @@ object API {
     println(s"Server API expected at: $apiUrl")
     
     implicit val sttpBackend: SttpBackend[concurrent.Future,_] = sttp.client3.FetchBackend()
-    val client = new ApiClient.DefaultImpl(ApiClient.Config(apiUrl))
+    val client = new ApiClient(ApiClient.Config(apiUrl))
     val admin = new AdminDataExplorerApiClient.DefaultImpl(AdminDataExplorerApiClient.Config(apiUrl))
     val adminApi = APIAdmin(admin, apiUrl)
     API(client, adminApi)
