+++ web/src/main/scala/net/wiringbits/API.scala
@@ -1,13 +1,13 @@
 package net.wiringbits
 
-import net.wiringbits.api.{ApiClient, SttpTapirApiClient}
+import net.wiringbits.api.ApiClient
 import net.wiringbits.services.StorageService
 import org.scalajs.macrotaskexecutor.MacrotaskExecutor.Implicits.*
 import sttp.client3.SttpBackend
 
 import scala.concurrent.Future
 
-case class API(client: ApiClient, storage: StorageService, tapirClient: SttpTapirApiClient)
+case class API(client: ApiClient, storage: StorageService)
 
 object API {
 
@@ -22,10 +22,9 @@ object API {
     println(s"Server API expected at: $apiUrl")
 
     implicit val sttpBackend: SttpBackend[Future, _] = sttp.client3.FetchBackend()
-    val tapirClient = new SttpTapirApiClient(SttpTapirApiClient.Config(apiUrl))
-    val client = new ApiClient.DefaultImpl(ApiClient.Config(apiUrl))
+    val client = new ApiClient(ApiClient.Config(apiUrl))
     val storage = new StorageService
 
-    API(client, storage, tapirClient)
+    API(client, storage)
   }
 }
