+++ web/src/main/scala/net/wiringbits/AppRouter.scala
@@ -51,7 +51,7 @@ object AppRouter {
       DashboardPage.Props(props.ctx, user)))
     def me(user: User) = route("/me", props.ctx)(UserEditPage.component(UserEditPage.Props(props.ctx, user)))
     val signOut = route("/signout", props.ctx) {
-      props.ctx.api.tapirClient.logout.onComplete {
+      props.ctx.api.client.logout.onComplete {
         case Success(_) =>
           props.ctx.loggedOut()
           println("Logged out successfully")
