+++ web/src/main/scala/net/wiringbits/components/widgets/EditPasswordForm.scala
@@ -50,7 +50,7 @@ object EditPasswordForm {
               setFormData(_.submissionFailed(texts.completeData))
               None
             }
-        } yield props.ctx.api.tapirClient
+        } yield props.ctx.api.client
           .updatePassword(request)
           .onComplete {
             case Success(_) =>
