+++ web/src/main/scala/net/wiringbits/components/widgets/ForgotPasswordForm.scala
@@ -50,7 +50,7 @@ object ForgotPasswordForm {
               setFormData(_.submissionFailed(texts.completeData))
               None
             }
-        } yield props.ctx.api.tapirClient
+        } yield props.ctx.api.client
           .forgotPassword(request)
           .onComplete {
             case Success(_) =>
