+++ web/src/main/scala/net/wiringbits/components/widgets/ResetPasswordForm.scala
@@ -54,7 +54,7 @@ object ResetPasswordForm {
               setFormData(_.submissionFailed(texts.completeData))
               None
             }
-        } yield props.ctx.api.tapirClient
+        } yield props.ctx.api.client
           .resetPassword(request)
           .onComplete {
             case Success(res) =>
