+++ web/src/main/scala/net/wiringbits/components/widgets/SignUpForm.scala
@@ -54,7 +54,7 @@ object SignUpForm {
               setFormData(_.submissionFailed(texts.completeData))
               None
             }
-        } yield props.ctx.api.tapirClient
+        } yield props.ctx.api.client
           .createUser(request)
           .onComplete {
             case Success(_) =>
